Move AdvancementMode enum to page-advancement.js and add utilities
